Podcast_Mini_Backend/requirements.txt
Django>=3.2
django-cors-headers>=3.7.0
djangorestframework>=3.12.4
Pillow>=8.2.0
python-dotenv>=0.17.0
